Mutualisation de code des serveurs sur l'endpoint qui sert les propositions de trajets + ne pas considérer qu'il y a un trajet retour quand il n'y a pas de retour Fixes #97
